requirements.txt
django
django-filter
djangorestframework
djangorestframework-jsonapi
psycopg2
channels
channels_redis
websocket_client
asyncio
django-cors-headers
python-dotenv
requests
django-allauth
dj-rest-auth
asgiref
djangorestframework-simplejwt
cryptography
newrelic